refactor(backend): divide login
Dividi as responsabilidades do login em `auth.service.js` e `auth.controller.js` ─ Também altera os arquivos `.repositories.js` para `.repository.js` e cria as pastas `services` e `controllers`
