+++ projeto/src/controllers/auth.controller.js
@@ -1,0 +1,22 @@
+const { loginService } = require("../services/auth.service");
+
+async function loginController(req, res) {
+  const { cpf, senha } = req.body;
+
+  if (!cpf || !senha) {
+    return res.status(400).json({ message: "CPF e senha são obrigatórios" });
+  }
+
+  try {
+    const result = await loginService(cpf,senha);
+    return res.status(200).json(result);
+  } catch (e) {
+    return res.status(500).json({
+      message: e.message,
+    });
+  }
+}
+
+module.exports = {
+  loginController,
+};
