+++ projeto/src/routes/auth.routes.js
@@ -1,31 +1,9 @@
 const { Router } = require("express");
-const {
-  findUsuarioByCpfAndSenha,
-} = require("../repositories/usuarios.repositories");
-const { createToken } = require("../utils/jwt");
+const { loginController } = require("../controllers/auth.controller");
 
 const router = Router();
 
-router.post("/login", async function(req, res) {
-  const { cpf, senha } = req.body;
-
-  if (!cpf || !senha) {
-    return res.status(400).json({ message: "CPF e senha são obrigatórios" });
-  }
-
-  try {
-    const usuario = await findUsuarioByCpfAndSenha(cpf, senha);
-    const token = createToken({id_usuario:usuario.id_usuario});
-    return res.status(200).json({
-      token,
-      nome: usuario.nome,
-    });
-  } catch (e) {
-    return res.status(500).json({
-      message: e.message,
-    });
-  }
-});
+router.post("/login", loginController);
 
 module.exports = router;
